fix: publish revision in service
