+++ src/modules/harvest/harvest.service.ts
@@ -35,7 +35,7 @@ export class HarvestService {
 
     if (!harvest) {
       throw new HttpException(
-        `Source ${harvestId} not found`,
+        `Harvest ${harvestId} not found`,
         HttpStatus.NOT_FOUND,
       );
     }
